Выносим render в отдельный модуль
